refactor(executor): split columnar_execution.rs into focused modules
Split the 1364-line columnar_execution.rs file into 5 focused submodules:
- mod.rs (392 lines) - Main entry points and orchestration
- group_by.rs (320 lines) - GROUP BY execution with CSE optimization
- join.rs (344 lines) - Multi-table JOIN execution
- join_helpers.rs (250 lines) - JOIN helper functions
- cse.rs (122 lines) - Common Sub-Expression Elimination helpers

All modules are under 400 lines as per acceptance criteria.
No functional changes - pure refactoring.

Closes #3290
